+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+Nothing yet
+
+## [v0.8.0] - 2021-04-12
+
 ### Added
 
 - You can now call `network.db.getWaveforms` to load waveforms for a track.
@@ -150,7 +154,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial working implementation. This is currently being used to re-implement
   [prolink-tools](https://github.com/evanpurkhiser/prolink-tools).
 
-[Unreleased]: https://github.com/evanpurkhiser/prolink-connect/compare/v0.7.2...HEAD
+[Unreleased]: https://github.com/evanpurkhiser/prolink-connect/compare/v0.8.0...HEAD
+[v0.8.0]: https://github.com/evanpurkhiser/prolink-connect/compare/v0.7.2...v0.8.0
 [v0.7.2]: https://github.com/evanpurkhiser/prolink-connect/compare/v0.7.1...v0.7.2
 [v0.7.1]: https://github.com/evanpurkhiser/prolink-connect/compare/v0.7.0...v0.7.1
 [v0.7.0]: https://github.com/evanpurkhiser/prolink-connect/compare/v0.6.0...v0.7.0